Saves time by quarantining only the dirtiest tests
